+++ app/lib/frontend/handlers/routes.g.dart
@@ -47,12 +47,9 @@ Router _$PubSiteServiceRouter(PubSiteService service) {
   router.add('GET', r'/favicon.ico', service.staticAsset);
   router.add('GET', r'/static/<path|[^]*>', service.staticAsset);
   router.add('GET', r'/experimental', service.experimental);
-  router.add('GET', r'/oauth/callback', service.oauthCallback);
   router.add('GET', r'/my-packages', service.accountPackagesPage);
   router.add('GET', r'/my-publishers', service.accountPublishersPage);
   router.add('GET', r'/authorized', service.authorizationConfirmed);
-  router.add('GET', r'/admin/confirm/new-uploader/<package>/<email>/<nonce>',
-      service.confirmUploader);
   router.add('GET', r'/consent', service.consentPage);
   return router;
 }