+++ app/lib/frontend/templates/package_versions.dart
@@ -124,7 +124,7 @@ String renderPkgVersionsPage(
     title: '${package.name} package - All Versions',
     canonicalUrl: urls.pkgPageUrl(package.name, includeHost: true),
     pageData: pkgPageData(package, latestVersion),
-    noIndex: package.isDiscontinued ?? false,
+    noIndex: package.isDiscontinued,
   );
 }