+++ app/lib/shared/user_merger.dart
@@ -140,14 +140,22 @@ class UserMerger {
       },
     );
 
-    // Consent of the User
+    // Consent's userId attribute
     await _processConcurrently(
-      _db.query<Consent>(ancestorKey: fromUserKey),
+      _db.query<Consent>()..filter('userId =', fromUserId),
       (Consent m) async {
+        if (m?.parentKey?.id != null) {
+          throw StateError('Old Consent entity: ${m.consentId}.');
+        }
         await _db.withTransaction((tx) async {
-          tx.queueMutations(
-              inserts: [m.changeParentUserId(toUserId)], deletes: [m.key]);
-          await tx.commit();
+          final consent = await tx.lookupValue<Consent>(m.key);
+          if (consent.userId == fromUserId) {
+            consent.userId = toUserId;
+            tx.queueMutations(inserts: [consent]);
+            await tx.commit();
+          } else {
+            await tx.rollback();
+          }
         });
       },
     );
@@ -156,6 +164,9 @@ class UserMerger {
     await _processConcurrently(
       _db.query<Consent>()..filter('fromUserId =', fromUserId),
       (Consent m) async {
+        if (m?.parentKey?.id != null) {
+          throw StateError('Old Consent entity: ${m.consentId}.');
+        }
         await _db.withTransaction((tx) async {
           final consent = await tx.lookupValue<Consent>(m.key);
           if (consent.fromUserId == fromUserId) {